+++ tests/testthat/test-workStudy.R
@@ -56,8 +56,9 @@ test_that('Test working with AcousticStudy object', {
     expect_true(all(c('Click_Detector_1_ici', 'All_ici') %in% names(getMeasures(exData[[1]]))))
     expect_true(all(c('Click_Detector_1_ici', 'All_ici') %in% names(getClickData(exData[[1]]))))
     iciData <- getICI(exData, 'data')
-    expect_true(all(c('Click_Detector_1', 'All') %in% names(iciData[[1]])))
-    expect_identical(names(iciData), names(events(exData)))
+    expect_true(all(c('Click_Detector_1', 'All') %in% iciData$detectorName))
+    # expect_identical(names(iciData), names(events(exData)))
+    expect_true(all(names(events(exData)) %in% iciData$eventId))
     iciData <- getICI(exData, 'value')
     expect_true(all(c('Click_Detector_1_ici', 'All_ici') %in% names(iciData[[1]])))
 
@@ -213,4 +214,12 @@ test_that('Test updateFiles', {
     expect_true(!any(file.exists(files(exStudy)$recordings$file)))
     exStudy <- updateFiles(exStudy, recording=recs, verbose=FALSE)
     expect_true(all(file.exists(files(exStudy)$recordings$file)))
+})
+
+test_that('Test bindStudies', {
+    data(exStudy)
+    expect_warning(bind2 <- bindStudies(exStudy, exStudy), 'Duplicate names')
+    expect_equal(nClicks(exStudy)*2, nClicks(bind2))
+    bind2list <- expect_warning(bindStudies(list(exStudy, exStudy)))
+    expect_equal(nClicks(exStudy)*2, nClicks(bind2list))
 })